# apps/backend/docs/swagger.yaml
basePath: /
definitions:
...
  internal_handlers.CreateNotificationChannelRequest:
    properties:
      email:
        type: string
      fcmToken:
        type: string
      type:
        type: string
    type: object
  internal_handlers.CreateNotificationChannelResponse:
    properties:
      endpointArn:
        type: string
      message:
        type: string
    type: object
...
  internal_handlers.SendNotificationRequest:
    properties:
      email:
        type: string
      endpointArn:
        type: string
      message:
        type: string
      subject:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  title: My API Name
  version: "1.0"
paths:
...
  /api/notifications/channels:
    post:
      consumes:
      - application/json
      description: Sets up email or push notification channel for the user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Notification channel request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateNotificationChannelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.CreateNotificationChannelResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a notification channel
      tags:
      - notifications
  /api/notifications/test:
    post:
      consumes:
      - application/json
      description: Sends a test notification via email or push
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Send notification request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.SendNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send a test notification
      tags:
      - notifications
...
swagger: "2.0"